# apigw/openapi-with-backend.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Claim Status API",
    "version": "1.0.0"
  },
  "paths": {
    "/claims/health": {
      "get": {
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "x-amazon-apigateway-integration": {
          "type": "http_proxy",
          "httpMethod": "GET",
          "uri": "http://a88fb53cff9ab467aa258cf5d46e5c36-85482127.us-east-1.elb.amazonaws.com/claims/health",
          "passthroughBehavior": "when_no_match",
          "requestParameters": {},
          "responses": {
            "default": {
              "statusCode": "200"
            }
          }
        }
      }
    },
    "/claims/{id}": {
      "get": {
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "x-amazon-apigateway-integration": {
          "type": "http_proxy",
          "httpMethod": "GET",
          "uri": "http://a88fb53cff9ab467aa258cf5d46e5c36-85482127.us-east-1.elb.amazonaws.com/claims/{id}",
          "passthroughBehavior": "when_no_match",
          "requestParameters": {
            "integration.request.path.id": "method.request.path.id"
          },
          "responses": {
            "default": {
              "statusCode": "200"
            }
          }
        }
      }
    },
    "/claims/{id}/summarize": {
      "post": {
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "x-amazon-apigateway-integration": {
          "type": "http_proxy",
          "httpMethod": "POST",
          "uri": "http://a88fb53cff9ab467aa258cf5d46e5c36-85482127.us-east-1.elb.amazonaws.com/claims/{id}/summarize",
          "passthroughBehavior": "when_no_match",
          "requestParameters": {
            "integration.request.path.id": "method.request.path.id"
          },
          "responses": {
            "default": {
              "statusCode": "200"
            }
          }
        }
      }
    }
  }
}